adds phase selector to submissions, no longer hard coded
