+++ docker-compose.yml
@@ -77,7 +77,7 @@ services:
   #   Database Service
   #-----------------------------------------------
   db:
-    image: postgres
+    image: postgres:10.6-alpine
     container_name: db
     env_file: .env
     environment:
