+++ requirements.txt
@@ -11,7 +11,7 @@ pillow==5.0.0
 celery==4.2.1
 gunicorn==19.9.0
 flower==0.9.2
-uvicorn==0.3.13
+uvicorn==0.3.21
 #daphne==2.2.2
 pyyaml==3.13
 watchdog==0.8.3
@@ -38,6 +38,5 @@ pygments==2.2.0
 drf-writable-nested==0.4.1
 django-filter==2.0.0
 drf-yasg[validation]==1.11.0
-flex=6.12.0
-Click=7.0
-
+flex==6.12.0
+Click==7.0
