+++ src/apps/api/serializers/competitions.py
@@ -50,52 +50,6 @@ class PageSerializer(WritableNestedModelSerializer):
         )
 
 
-class SubmissionSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Submission
-        fields = (
-            'phase',
-            'name',
-            'description',
-            'pk',
-            'id',
-            'created_when',
-            'is_public',
-            'status',
-            'status_details',
-            'secret',
-        )
-        extra_kwargs = {
-            "secret": {
-                "write_only": True
-            }
-        }
-
-    def update(self, instance, validated_data):
-        if instance.secret != validated_data.get('secret'):
-            raise PermissionError("Submission secret invalid")
-        return super().update(instance, validated_data)
-
-
-class SubmissionCreationSerializer(serializers.ModelSerializer):
-    data = serializers.SlugRelatedField(queryset=Data.objects.all(), required=False, allow_null=True, slug_field='key')
-
-    class Meta:
-        model = Submission
-        fields = (
-            'data',
-            'phase',
-        )
-
-    # TODO: Validate the user is a participant in this competition.phase
-
-    def create(self, validated_data):
-        validated_data["owner"] = self.context['owner']
-        sub = super().create(validated_data)
-        sub.start()
-        return sub
-
-
 class CompetitionSerializer(WritableNestedModelSerializer):
     created_by = serializers.SerializerMethodField(read_only=True)
     logo = NamedBase64ImageField(required=True)
