+++ src/apps/api/urls.py
@@ -5,14 +5,14 @@ from drf_yasg.views import get_schema_view
 from rest_framework.routers import SimpleRouter
 from rest_framework.permissions import AllowAny
 
-from .views import competitions, datasets, profiles, leaderboards
+from .views import competitions, datasets, profiles, leaderboards, submissions
 
 
 router = SimpleRouter()
 router.register('competitions', competitions.CompetitionViewSet)
 router.register('competition_status', competitions.CompetitionCreationTaskStatusViewSet)
 router.register('phases', competitions.PhaseViewSet)
-router.register('submissions', competitions.SubmissionViewSet)
+router.register('submissions', submissions.SubmissionViewSet)
 router.register('datasets', datasets.DataViewSet)
 router.register('data_groups', datasets.DataGroupViewSet)
 router.register('leaderboards', leaderboards.LeaderboardViewSet)
