+++ src/apps/competitions/views.py
@@ -1,5 +1,5 @@
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.views.generic import TemplateView
+from django.views.generic import TemplateView, DetailView
 from django.views.generic.base import ContextMixin
 
 from .models import Competition
@@ -17,17 +17,18 @@ class CompetitionUpload(LoginRequiredMixin, TemplateView):
     template_name = 'competitions/upload.html'
 
 
-class CompetitionDetail(LoginRequiredMixin, TemplateView, ContextMixin):
+class CompetitionDetail(LoginRequiredMixin, DetailView):
+    queryset = Competition.objects.all()
     template_name = 'competitions/detail.html'
 
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        # context['number'] = random.randrange(1, 100)
-        try:
-            my_pk = kwargs.pop('pk')
-            comp = Competition.objects.get(pk=my_pk)
-            context['comp'] = comp
-            context['comp_pk'] = comp.pk
-        except KeyError:
-            print("Key error oh no")
-        return context
+    # def get_context_data(self, **kwargs):
+    #     context = super().get_context_data(**kwargs)
+    #     # # context['number'] = random.randrange(1, 100)
+    #     # try:
+    #     #     my_pk = kwargs.pop('pk')
+    #     #     comp = Competition.objects.get(pk=my_pk)
+    #     #     context['comp'] = comp
+    #     #     context['comp_pk'] = comp.pk
+    #     # except KeyError:
+    #     #     print("Key error oh no")
+    #     return context
