--- src/static/riot/competitions/detail/_run_info.tag
@@ -1,139 +1,0 @@
-<comp-run-info>
-    <div class="ui container grid comp-run-info">
-        <div class="centered row">
-            <div class="five wide column">
-                <h4 class="ui top attached header">
-                    Leaderboard
-                </h4>
-                <div class="ui attached message forced-full-height">
-                    <div class="ui relaxed divided mini list">
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">Participant</div>
-                                Metric 1: x Metric 2: x
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">Participant</div>
-                                Metric 1: x Metric 2: x
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">Participant</div>
-                                Metric 1: x Metric 2: x
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">Participant</div>
-                                Metric 1: x Metric 2: x
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">Participant</div>
-                                Metric 1: x Metric 2: x
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">Participant</div>
-                                Metric 1: x Metric 2: x
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="five wide column">
-                <h4 class="ui top attached header">
-                    Public Submissions
-                </h4>
-                <div class="ui attached message forced-full-height">
-                    <div class="ui relaxed divided mini list">
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">
-                                    <div class="item">Submission_01 by (user)</div>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">
-                                    <div class="item">Submission_01 by (user)</div>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">
-                                    <div class="item">Submission_01 by (user)</div>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">
-                                    <div class="item">Submission_01 by (user)</div>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">
-                                    <div class="item">Submission_01 by (user)</div>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">
-                                    <div class="item">Submission_01 by (user)</div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="five wide column">
-                <h4 class="ui top attached header">
-                    Comment Feed
-                </h4>
-                <div class="ui attached message forced-full-height">
-                    <div class="ui relaxed divided list">
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">Snickerdoodle</div>
-                                An excellent companion
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">Poodle</div>
-                                A poodle, its pretty basic
-                            </div>
-                        </div>
-                        <div class="item">
-                            <div class="content">
-                                <div class="header">Paulo</div>
-                                He's also a dog
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    <div class="ui horizontal divider"></div>
-    <style type="text/stylus">
-        .comp-run-info
-            /*padding 50vh !important*/
-            margin 5vh !important
-    </style>
-    <script>
-        var self = this
-        self.competition = opts.competition
-    </script>
-</comp-run-info>
