--- src/static/riot/competitions/detail/_tags.tag
@@ -1,23 +1,0 @@
-<comp-tags>
-    <div class="ui container relaxed grid">
-        <div align="center" class="row">
-            <div class="sixteen wide column">
-                <div class="ui">
-                    <h1>Tags</h1>
-                    <div class="ui tag labels">
-                        <a class="ui label">
-                            <i class="icon settings"></i> Beginner
-                        </a>
-                        <a class="ui label">
-                            <i class="icon wrench"></i> Mechanics
-                        </a>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    <script>
-        var self = this
-        self.competition = opts.competition
-    </script>
-</comp-tags>
